+++ CHANGELOG.md
@@ -82,3 +82,4 @@
 - Added error log.
 - Added specific refresh rate for error log.
 - Added include and exclude filters to error log.
+- Added linechart to error log.